added package_cmd
