+++ fedn/cli/__init__.py
@@ -1,2 +1,3 @@
 from .main import main  # noqa: F401
 from .run_cmd import run_cmd  # noqa: F401
+from .package_cmd import package_cmd  # noqa: F401
